+++ tests/src/generator.test.ts
@@ -1,10 +1,10 @@
 import { readFileSync } from "fs";
 import { GraphQLSchema } from "graphql";
 import path from "path";
-import { getMocks } from "../generator";
+import { getMocks } from "../../src/generator";
 
 const schema = readFileSync(
-  path.join(__dirname, "./resources/testSchema.graphql"),
+  path.join(__dirname, "../resources/testSchema.graphql"),
   "utf8"
 ) as unknown as GraphQLSchema;
