+++ tests/src/utils/state/stateMachine.test.ts
@@ -1,7 +1,7 @@
 import {
   getEntityInstance,
   StateMachine,
-} from "../../../utils/state/stateMachine";
+} from "../../../../src/utils/state/stateMachine";
 
 const { statesData, stateMachine } = {
   statesData: {
